+++ exercises/exercise_sheet_4/4_8_2.py
@@ -2,23 +2,23 @@ import numpy as np
 from statistics import covariance as cov
 
 
-def expected_value(values: list) -> int:
+def expected_value(values: list) -> float:
     return np.average(values)
 
 
-def variance(values: list) -> int:
+def variance(values: list) -> float:
     return np.var(values)
 
 
-def standard_deviation(values: list) -> int:
+def standard_deviation(values: list) -> float:
     return np.std(values)
 
 
-def covariance(values1: list, values2: list) -> int:
+def covariance(values1: list, values2: list) -> float:
     return cov(values1, values2)
 
 
-def correlation_coefficient(values1: list, values2: list) -> int:
+def correlation_coefficient(values1: list, values2: list) -> float:
     return covariance(values1, values2) / (standard_deviation(values1) * standard_deviation(values2))
